Remove automatic URL updates
Now that we bundle local version of the files there is no need to
update the URLs for downloading the installer. This improvement is
only possible because OpenVPN has an unchanging latest release URL
that can be used instead of the explicitly versioned URLs.

The old way, using the explicitly versioned URLs, is preferable if the
files aren't bundled. That is because if the file the URL points to on
openvpn.net is updated, but the chocolatey package has not yet been
updated, then the checksums will not match and user installs will
fail. The files are now being downloaded from chocolatey.org rather
than openvpn.net so unupdated checksums should not be an issue.